Split JUPR Live into public and admin pages
